pyproject.toml
[project]
name = "anyformat"
version = "0.1.0"
description = "CLI tool to convert media files between formats"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.13"
authors = [{ name = "Open Source Workshop" }]
keywords = ["cli", "media", "converter", "ffmpeg", "audio", "video", "image"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Multimedia :: Video :: Conversion",
    "Topic :: Multimedia :: Sound/Audio :: Conversion",
    "Topic :: Utilities",
]
dependencies = [
    "typer>=0.23.0",
    "rich>=14.0.0",
    "pillow>=12.0.0",
    "pydub>=0.25.0",
    "ffmpeg-python>=0.2.0",
    "audioop-lts>=0.2.0; python_version >= '3.13'",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.8.0",
]

[project.scripts]
anyformat = "anyformat.cli:app"

[project.urls]
Homepage = "https://github.com/anxkhn/anyformat-workshop"
Repository = "https://github.com/anxkhn/anyformat-workshop"
Issues = "https://github.com/anxkhn/anyformat-workshop/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/anyformat"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.ruff]
target-version = "py313"
line-length = 100
exclude = [".git", "__pycache__", ".venv", "build", "dist"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["anyformat"]

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]